requirements.txt
streamlit 
langchain 
openai 
faiss-cpu 
PyPDF2 
python-dotenv 
langchain-community